indicator.html embed test
It’s definitely wrong, but as a starting point how would I edit this Javascript to refresh the embedmap iframe at the end of the indicator.html?

{% if meta.embedded_map_html %}
  {% meta.getElementById('embedded_map_html.id').src = document.getElementById('embedded_map_html.id').src %}
{% endif %}

I’ve added it to the end of the file in develop, so it’ll override the remote in the sandpit site. Could you show me please how to link it to the iframe id or name (perhaps calling it through a ‘name=’ function in the iframe which relates to the indicator code)?