Push to master (#3022)
Took Savvas' code and got something working on the UK site. However, this is  a short-term fix as I eventually just hardcoded the link and commented out the survey tags. This is because it displays on a new line that obscures the headers, even when a similar formatting to the 'dev disclaimer' is applied.

No luck pulling across from open-sdg, but it is possible. All local for now.

I added the following updates to sdg-indicators...
https://github.com/ONSdigital/sdg-indicators/blob/develop/data/translations/en/header.yml
https://github.com/ONSdigital/sdg-indicators/blob/develop/_includes/components/survey-message.html
https://github.com/ONSdigital/sdg-indicators/blob/develop/_includes/header.html
https://github.com/ONSdigital/sdg-indicators/blob/02af4a2fc55591d6fee59dbb7b22db4b954b2c0a/_config.yml#L156